Day 7
The problem was fun, and I was absolutely stupid in part2.
Part 1 was a slightly slow affair because I misunderstood what the question was asking. But once I got the correct answer for the test input, it also worked on the large input without any corrections.

Part 2 though, oh boy. Initially I thought of a DFS style approach. Then, I didn't want to use a tree style data structure, so I instead elected to use sets. But for a 150+ line input and a final answer that is 1.5 trillion+, who knew creating set elements would be too slow??!!

Anyways, I then realised that sets are not needed, because every path down is unique, as opposed to part 1 where every beam is non unique w.r.t how many splitters contribute to it.

Long story short, use an array of numbers and bam! 0.01 seconds runtime for entire part1 + part 2