Day 8 code
Where do I even begin?

So, looking at the problem, I first think of a heap-based solution. Part 1 was straightforward enough.

For part 2, my initial solution was an inefficient O(n) version of UnionFind. The first solution I generated and got accepted, the runtime was ~1.16 seconds.

I was very unhappy with this, so I implemented the following changes:
1. Improve the heap construction to be O(n) by heapifying the entire array at once, rather than O(nlogn) adding edges 1 by 1
2. Implement a Union Find class to do it properly and efficiently.
2a. Add a component counter to the UF class, so that I do not need any more calculations to figure out when the graph is fully connected.
3. Flatten junction storage into 3 lists rather than individual tuples.

With all these changes, final runtime was ~0.4 seconds. By far the worst this AOC, but still sub 0.5, so it will do for now.