Day 11
Simple BFS was more than enough for part 1. Part 2 went nuts with a BFS.

I thought of using a DFS for part 2, but then accidentally realized that the graph was a DAG, so I decided to learn topological sort and use that to count paths.

The code is much longer than a simple memoized DFS could do, but the runtime is similar, so I don't really care.

There are 2 small input files because they are different in the problem.